Переместил проект Дз 5-6 из основной папки Homework в модуль task5_6
