--- Homework/src/main/java/org/yakdanol/homework/annotation/LogExecutionTime.java
@@ -1,11 +1,0 @@
-package org.yakdanol.homework.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target({ElementType.METHOD, ElementType.TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface LogExecutionTime {
-}
