--- Homework/src/main/java/org/yakdanol/homework/aspect/LogAspect.java
@@ -1,23 +1,0 @@
-package org.yakdanol.homework.aspect;
-
-import lombok.extern.slf4j.Slf4j;
-import org.aspectj.lang.ProceedingJoinPoint;
-import org.aspectj.lang.annotation.Around;
-import org.aspectj.lang.annotation.Aspect;
-import org.springframework.stereotype.Component;
-
-@Aspect
-@Component
-@Slf4j
-public class LogAspect {
-
-    @Around("@annotation(org.yakdanol.homework.annotation.LogExecutionTime) || @within(org.yakdanol.homework.annotation.LogExecutionTime)")
-    public Object logExecutionTime(ProceedingJoinPoint joinPoint) throws Throwable {
-        long start = System.currentTimeMillis();
-        Object proceed = joinPoint.proceed();
-        long executionTime = System.currentTimeMillis() - start;
-        log.info("{} executed in {} ms", joinPoint.getSignature(), executionTime);
-
-        return proceed;
-    }
-}
