--- Homework/src/main/java/org/yakdanol/homework/controller/CategoryController.java
@@ -1,45 +1,0 @@
-package org.yakdanol.homework.controller;
-
-import org.yakdanol.homework.annotation.LogExecutionTime;
-import org.yakdanol.homework.model.Category;
-import org.yakdanol.homework.service.CategoryService;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.concurrent.ConcurrentMap;
-
-@RestController
-@RequestMapping("/api/v1/places/categories")
-@LogExecutionTime
-public class CategoryController {
-
-    private final CategoryService categoryService;
-
-    public CategoryController(CategoryService categoryService) {
-        this.categoryService = categoryService;
-    }
-
-    @GetMapping
-    public ConcurrentMap<Long, Category> getAllCategories() {
-        return categoryService.getAllCategories();
-    }
-
-    @GetMapping("/{id}")
-    public Category getCategoryById(@PathVariable Long id) {
-        return categoryService.getCategoryById(id);
-    }
-
-    @PostMapping
-    public Category createCategory(@RequestBody Category category) {
-        return categoryService.createCategory(category.getId(), category);
-    }
-
-    @PutMapping("/{id}")
-    public Category updateCategory(@PathVariable Long id, @RequestBody Category category) {
-        return categoryService.updateCategory(id, category);
-    }
-
-    @DeleteMapping("/{id}")
-    public void deleteCategory(@PathVariable Long id) {
-        categoryService.deleteCategory(id);
-    }
-}
