--- Homework/src/main/java/org/yakdanol/homework/exception/CategoryNotFoundException.java
@@ -1,7 +1,0 @@
-package org.yakdanol.homework.exception;
-
-public class CategoryNotFoundException extends RuntimeException {
-    public CategoryNotFoundException(Long id) {
-        super("Category not found with ID: " + id);
-    }
-}
