--- Homework/src/main/java/org/yakdanol/homework/exception/EmptyRepositoryException.java
@@ -1,7 +1,0 @@
-package org.yakdanol.homework.exception;
-
-public class EmptyRepositoryException extends RuntimeException {
-    public EmptyRepositoryException(String repositoryName) {
-        super(repositoryName + " is empty.");
-    }
-}
