--- Homework/src/main/java/org/yakdanol/homework/exception/GlobalExceptionHandler.java
@@ -1,25 +1,0 @@
-package org.yakdanol.homework.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(CategoryNotFoundException.class)
-    public ResponseEntity<String> handleCategoryNotFound(CategoryNotFoundException ex) {
-        return new ResponseEntity<>(ex.getMessage(), HttpStatus.NOT_FOUND);
-    }
-
-    @ExceptionHandler(LocationNotFoundException.class)
-    public ResponseEntity<String> handleLocationNotFound(LocationNotFoundException ex) {
-        return new ResponseEntity<>(ex.getMessage(), HttpStatus.NOT_FOUND);
-    }
-
-    @ExceptionHandler(EmptyRepositoryException.class)
-    public ResponseEntity<String> handleEmptyRepository(EmptyRepositoryException ex) {
-        return new ResponseEntity<>(ex.getMessage(), HttpStatus.NO_CONTENT);
-    }
-}
