--- Homework/src/main/java/org/yakdanol/homework/exception/LocationNotFoundException.java
@@ -1,7 +1,0 @@
-package org.yakdanol.homework.exception;
-
-public class LocationNotFoundException extends RuntimeException {
-    public LocationNotFoundException(String slug) {
-        super("Location not found with slug: " + slug);
-    }
-}
