--- Homework/src/main/java/org/yakdanol/homework/model/Category.java
@@ -1,16 +1,0 @@
-package org.yakdanol.homework.model;
-
-import lombok.*;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
-public class Category {
-    private Long id;
-
-    @NonNull
-    private String slug;
-
-    @NonNull
-    private String name;
-}
