--- Homework/src/main/java/org/yakdanol/homework/model/Location.java
@@ -1,17 +1,0 @@
-package org.yakdanol.homework.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.NonNull;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
-public class Location {
-    @NonNull
-    private String slug;
-
-    @NonNull
-    private String name;
-}
