+++ Homework/task5_6/src/main/java/org/yakdanol/task5_6/Task56Application.java
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class Task56Application {
 
-    public static void main(String[] args) {
-        SpringApplication.run(Task56Application.class, args);
-    }
+	public static void main(String[] args) {
+		SpringApplication.run(Task56Application.class, args);
+	}
 
 }
