+++ Homework/task5_6/src/main/java/org/yakdanol/task5_6/annotation/LogExecutionTime.java
@@ -1,0 +1,11 @@
+package org.yakdanol.task5_6.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target({ElementType.METHOD, ElementType.TYPE})
+@Retention(RetentionPolicy.RUNTIME)
+public @interface LogExecutionTime {
+}
