+++ Homework/task5_6/src/main/java/org/yakdanol/task5_6/aspect/LogAspect.java
@@ -1,0 +1,23 @@
+package org.yakdanol.task5_6.aspect;
+
+import lombok.extern.slf4j.Slf4j;
+import org.aspectj.lang.ProceedingJoinPoint;
+import org.aspectj.lang.annotation.Around;
+import org.aspectj.lang.annotation.Aspect;
+import org.springframework.stereotype.Component;
+
+@Aspect
+@Component
+@Slf4j
+public class LogAspect {
+
+    @Around("@annotation(org.yakdanol.task5_6.annotation.LogExecutionTime) || @within(org.yakdanol.task5_6.annotation.LogExecutionTime)")
+    public Object logExecutionTime(ProceedingJoinPoint joinPoint) throws Throwable {
+        long start = System.currentTimeMillis();
+        Object proceed = joinPoint.proceed();
+        long executionTime = System.currentTimeMillis() - start;
+        log.info("{} executed in {} ms", joinPoint.getSignature(), executionTime);
+
+        return proceed;
+    }
+}
