+++ Homework/task5_6/src/main/java/org/yakdanol/task5_6/controller/CategoryController.java
@@ -1,0 +1,45 @@
+package org.yakdanol.task5_6.controller;
+
+import org.yakdanol.task5_6.annotation.LogExecutionTime;
+import org.yakdanol.task5_6.model.Category;
+import org.yakdanol.task5_6.service.CategoryService;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.concurrent.ConcurrentMap;
+
+@RestController
+@RequestMapping("/api/v1/places/categories")
+@LogExecutionTime
+public class CategoryController {
+
+    private final CategoryService categoryService;
+
+    public CategoryController(CategoryService categoryService) {
+        this.categoryService = categoryService;
+    }
+
+    @GetMapping
+    public ConcurrentMap<Long, Category> getAllCategories() {
+        return categoryService.getAllCategories();
+    }
+
+    @GetMapping("/{id}")
+    public Category getCategoryById(@PathVariable Long id) {
+        return categoryService.getCategoryById(id);
+    }
+
+    @PostMapping
+    public Category createCategory(@RequestBody Category category) {
+        return categoryService.createCategory(category.getId(), category);
+    }
+
+    @PutMapping("/{id}")
+    public Category updateCategory(@PathVariable Long id, @RequestBody Category category) {
+        return categoryService.updateCategory(id, category);
+    }
+
+    @DeleteMapping("/{id}")
+    public void deleteCategory(@PathVariable Long id) {
+        categoryService.deleteCategory(id);
+    }
+}
