+++ Homework/task5_6/src/main/java/org/yakdanol/task5_6/exception/CategoryNotFoundException.java
@@ -1,0 +1,7 @@
+package org.yakdanol.task5_6.exception;
+
+public class CategoryNotFoundException extends RuntimeException {
+    public CategoryNotFoundException(Long id) {
+        super("Category not found with ID: " + id);
+    }
+}
