+++ Homework/task5_6/src/main/java/org/yakdanol/task5_6/exception/EmptyRepositoryException.java
@@ -1,0 +1,7 @@
+package org.yakdanol.task5_6.exception;
+
+public class EmptyRepositoryException extends RuntimeException {
+    public EmptyRepositoryException(String repositoryName) {
+        super(repositoryName + " is empty.");
+    }
+}
