+++ Homework/task5_6/src/main/java/org/yakdanol/task5_6/exception/LocationNotFoundException.java
@@ -1,0 +1,7 @@
+package org.yakdanol.task5_6.exception;
+
+public class LocationNotFoundException extends RuntimeException {
+    public LocationNotFoundException(String slug) {
+        super("Location not found with slug: " + slug);
+    }
+}
