+++ Homework/task5_6/src/main/java/org/yakdanol/task5_6/model/Location.java
@@ -1,0 +1,17 @@
+package org.yakdanol.task5_6.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.NonNull;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Data
+public class Location {
+    @NonNull
+    private String slug;
+
+    @NonNull
+    private String name;
+}
